+++ pygeos/constructive.py
@@ -21,6 +21,7 @@ __all__ = [
     "make_valid",
     "normalize",
     "point_on_surface",
+    "polygonize",
     "reverse",
     "simplify",
     "snap",
@@ -484,6 +485,54 @@ def point_on_surface(geometry, **kwargs):
     <pygeos.Geometry POINT EMPTY>
     """
     return lib.point_on_surface(geometry, **kwargs)
+
+
+def polygonize(geometries, **kwargs):
+    """Creates polygons formed from the linework of a set of Geometries.
+
+    Polygonizes an array of Geometries that contain linework which
+    represents the edges of a planar graph. Any type of Geometry may be
+    provided as input; only the constituent lines and rings will be used to
+    create the output polygons.
+
+    Lines or rings that when combined do not completely close a polygon
+    will result in an empty GeometryCollection.  Duplicate segments are
+    ignored.
+
+    This function returns the polygons within a GeometryCollection.
+    Individual Polygons can be obtained using `get_geometry` to get
+    a single polygon or `get_parts` to get an array of polygons.
+    MultiPolygons can be constructed from the output using
+    ``pygeos.multipolygons(pygeos.get_parts(pygeos.polygonize(geometries)))``.
+
+    Parameters
+    ----------
+    geometries : array_like
+        An array of geometries.
+    axis : int
+        Axis along which the geometries are polygonized.
+        The default is to perform a reduction over the last dimension
+        of the input array. A 1D array results in a scalar geometry.
+
+    Returns
+    -------
+    GeometryCollection or array of GeometryCollections
+
+    See Also
+    --------
+    get_parts, get_geometry
+
+    Examples
+    --------
+    >>> lines = [
+    ...     Geometry("LINESTRING (0 0, 1 1)"),
+    ...     Geometry("LINESTRING (0 0, 0 1)"),
+    ...     Geometry("LINESTRING (0 1, 1 1)"),
+    ... ]
+    >>> polygonize(lines)
+    <pygeos.Geometry GEOMETRYCOLLECTION (POLYGON ((1 1, 0 0, 0 1, 1 1)))>
+    """
+    return lib.polygonize(geometries, **kwargs)
 
 
 @requires_geos("3.7.0")